Updated to react 0.9.0 with source map support
Adding full source map support requires a modified JSXTransform to be
used, as the Facebook JSXTransform does not seem to export the
transformCode method which handles the jsx processing as well as the
source map generation.

The transformCode also relies on the node.js buffer package, which
available using browserify within the transformer only.